service-saver/requirements.txt
grpcio==1.54.2
grpcio-tools==1.54.2
protobuf==4.23.2
SQLAlchemy==2.0.16
typing_extensions==4.6.3
